+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 Versions follow [Semantic Versioning](https://semver.org/spec/v2.0.0.html) (`<major>`.`<minor>`.`<patch>`)
 
 ## [vNext]
-* #7 Add support for in-block comments
+* #7 Add inline comment support
 
 ## [v0.2.0]
 ### Added